Fixing system maintenance page. Adding specs

+++ config/deploy.rb
@@ -2,19 +2,12 @@ require 'bundler/capistrano'
 require 'airbrake/capistrano'
 require 'capistrano-unicorn'
 require 'capistrano/maintenance'
-
 require './config/boot'
 
 default_run_options[:pty] = true
 #ssh_options[:port] = 31337
 ssh_options[:keys] = [File.join(ENV['HOME'], '.ssh', 'id_rsa')]
 ssh_options[:forward_agent] = true
-
-
-
-set :maintenance_template_path, 'app/views/maintenance.html.haml'
-set :maintenance_config_warning, false
-set :maintenance_dirname, "#{shared_path}/public/system"
 
 set :application, 'idealme'
 set :scm, :git
@@ -27,6 +20,14 @@ set :keep_releases, 5
 # Remove No such file/directory warnings.
 set :normalize_asset_timestamps, false
 
+set :maintenance_template_path, '/apps/idealme/current/app/views/maintenance.html.erb'
+set :maintenance_config_warning, false
+set :maintenance_dirname, '/apps/idealme/current/public/static'
+
+
+task :production do
+#...
+end
 
 task :staging do
   set :repository, 'ssh://git@bitbucket.org/billxinli/idealme.git'
@@ -35,32 +36,27 @@ task :staging do
   set :deploy_to, '/apps/idealme'
   set :branch, 'master'
   set :keep_releases, 1
+  set :db_destruction, true
+
   role :app, domain
   role :web, domain
   role :db, domain, :primary => true
 end
 
-
-namespace :deploy do
-  namespace :db do
-    desc 'drop database'
-    task :drop do
+namespace :db do
+  desc 'Drop database'
+  task :drop do
+    if exists?(:db_destruction)
       run "cd #{current_path}; bundle exec rake db:drop RAILS_ENV=#{rails_env}"
     end
-    desc 'reload the database with seed data'
-    task :seed do
+  end
+  
+  desc 'Seed database'
+  task :seed do
+    if exists?(:db_destruction)
       run "cd #{current_path}; bundle exec rake  db:seed:development:users db:seed:development:goals db:seed:development:goal_users db:seed:development:checkins db:seed:development:categories db:seed:development:jewels db:seed:development:courses RAILS_ENV=#{rails_env}"
     end
   end
-  #task :start do
-  #  ;
-  #end
-  #task :stop do
-  #  ;
-  #end
-  #task :restart, :roles => :app, :except => {:no_release => true} do
-  #  run "#{try_sudo} touch #{File.join(current_path, 'tmp', 'restart.txt')}"
-  #end
 end
 
 after 'deploy:restart', 'unicorn:restart' # app preloaded
